fix host serverd; more robust reconnect behavior

+++ script/host_serverd.py
@@ -2,7 +2,6 @@ import subprocess
 import asyncio
 import socket
 import traceback
-import shlex
 
 
 def check_slave_existence():
@@ -18,7 +17,8 @@ def check_slave_existence():
     proc.wait()
 
 
-def kill_slave(pid):
+def kill_proc(pid):
+    if pid is None: return 
     subprocess.run((rf'taskkill /F /PID {pid}'))
 
 
@@ -28,7 +28,7 @@ class RPyCSlaveDaemon():
         if pid is not None:
             print(
                 f'[INFO] Found slave running with pid {pid}. Killing it.')
-            kill_slave(pid)
+            kill_proc(pid)
         if port is None:
             port = 18816
         self.s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
@@ -39,38 +39,32 @@ class RPyCSlaveDaemon():
     async def handle(self, conn, port=None):
         if port is None:
             port = 18812
-        try:
-            with conn:
+        with conn:
+            while True:
                 data = (await self.loop.run_in_executor(None, lambda: conn.recv(100))).strip()
                 if data == b'start':
-                    proc = await asyncio.subprocess.create_subprocess_exec(
-                        shlex.split(rf"C:\Users\strontium_remote1\AppData\Local\Programs\Python\Python39\Python C:\Users\strontium_remote1\Desktop\launch_slave.py --host 0.0.0.0 -p {port}",
+                    await asyncio.subprocess.create_subprocess_shell(
+                        (rf"C:\Users\strontium_remote1\AppData\Local\Programs\Python\Python39\Python C:\Users\strontium_remote1\Desktop\launch_slave.py --host 0.0.0.0 -p {port}"),
                                     stdin=asyncio.subprocess.PIPE
-                                    ))
+                                    )
                     conn.sendall(b'done')
-                elif data == b'stop':
-                    proc.stdin.close()
-                    proc.kill()
-                    conn.sendall(b'done')
-                else:
-                    raise RuntimeError(f"unknown data {data}")
-        except asyncio.CancelledError:
-            proc.stdin.close()
-            proc.kill()
+                    return 
+                elif not len(data):
+                    return 
 
     async def serve(self):
         print('Waiting for connection...')
         task = None
         try:
-            conn, addr = await self.loop.run_in_executor(None, self.s.accept)
-            conn.sendall(b'good')
-            if task is not None:
-                if not task.cancelled() and not task.done():
-                    task.cancel()
-            task = asyncio.create_task(self.handle(conn))
+            while True:
+                conn, addr = await self.loop.run_in_executor(None, self.s.accept)
+                conn.sendall(b'good')
+                kill_proc(check_slave_existence())
+                task = asyncio.create_task(self.handle(conn))
         except Exception:
             traceback.print_exc()
         except:
+            traceback.print_exc()
             if not task.cancelled() and not task.done():
                 task.cancel()
